Moved protocols around, updated example(s)
