+++ examples/ethereum/list-stakes/main.go
@@ -6,22 +6,31 @@ package main
 
 import (
 	"context"
+	"encoding/json"
 	"errors"
+	"fmt"
 	"log"
 	"time"
 
+	"github.com/coinbase/staking-client-library-go/client/protocols"
 	"google.golang.org/api/iterator"
-	"google.golang.org/protobuf/encoding/protojson"
 
 	"github.com/coinbase/staking-client-library-go/auth"
 	"github.com/coinbase/staking-client-library-go/client"
 	"github.com/coinbase/staking-client-library-go/client/options"
-	"github.com/coinbase/staking-client-library-go/client/rewards"
-	"github.com/coinbase/staking-client-library-go/client/rewards/stakes"
-	rewardspb "github.com/coinbase/staking-client-library-go/gen/go/coinbase/staking/rewards/v1"
+	api "github.com/coinbase/staking-client-library-go/gen/go/coinbase/staking/rewards/v1"
 )
 
-// An example function to demonstrate how to use the staking client libraries.
+/*
+ * Run the code with 'go run main.go' to view the balances for the first validator on the Ethereum network.
+ * Or, to view balances for any arbitrary validator, simply replace the public key below.
+ */
+
+const (
+	// https://beaconcha.in/validator/1
+	address = "0xa1d1ad0714035353258038e964ae9675dc0252ee22cea896825c01458e1807bfad2f9969338798548d9858a571f7425c"
+)
+
 func main() {
 	ctx := context.Background()
 
@@ -38,16 +47,13 @@ func main() {
 		log.Fatalf("error instantiating staking client: %s", err.Error())
 	}
 
-	// List staking activities for a given protocol.
-	stakesIter := stakingClient.Rewards.ListStakes(ctx, &rewardspb.ListStakesRequest{
-		Parent:   rewards.Ethereum,
+	// List historical staking balances for the given address starting from the most recent after the given timestamp.
+	stakesIter := stakingClient.Rewards.ListStakes(ctx, &api.ListStakesRequest{
+		Parent:   protocols.Ethereum,
 		PageSize: 200,
-		Filter: stakes.WithAddress().Eq("0xac53512c39d0081ca4437c285305eb423f474e6153693c12fbba4a3df78bcaa3422b31d800c5bea71c1b017168a60474").
-			And(stakes.WithEvaluationTime().Lte(time.Now())).
-			String(),
+		Filter:   fmt.Sprintf(`address='%s' AND evaluation_time<='%s'`, address, time.Now().AddDate(0, 0, -20).Format(time.RFC3339)),
 	})
 
-	count := 0
 	for {
 		stake, err := stakesIter.Next()
 		if errors.Is(err, iterator.Done) {
@@ -58,12 +64,11 @@ func main() {
 			log.Fatalf("error listing stakes: %s", err.Error())
 		}
 
-		d, err := protojson.Marshal(stake)
+		marshaled, err := json.MarshalIndent(stake, "", "   ")
 		if err != nil {
-			log.Fatalf("error marshalling stake object: %s", err.Error())
+			log.Fatalf("error marshaling stake balance: %s", err.Error())
 		}
 
-		log.Printf("[%d] Stake details: %s", count, d)
-		count += 1
+		fmt.Println(string(marshaled))
 	}
 }
